[RENAME] 우선예약 관련 파일, 폴더 따로 만듦
