+++ flutter_front/lib/reservation_status/model/entity/pre_reservation/pre_reservation_status_entity.dart
