+++ flutter_front/lib/reservation_status/model/repository/pre_resevation/pre_reservation_setting_repository.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart' hide Headers;
 import 'package:flutter_front/common/dio/dio.dart';
-import 'package:flutter_front/reservation_status/model/entity/pre_reservation_status_entity.dart';
+import 'package:flutter_front/reservation_status/model/entity/pre_reservation/pre_reservation_status_entity.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:retrofit/retrofit.dart';
